Use the fast algorithm for floating point even in the presence of missing values, adapting existing code to handle `NaN`. After sorting `NaN` and `missing` at the end, a second pass is made over these to put missing after `NaN`.

Fixes https://github.com/JuliaLang/julia/issues/27781 (when combined with https://github.com/JuliaLang/julia/pull/27789).

I see a 2× performance improvement when the input contains 10% of missing values (and no `NaN`), making it about twice slower than a `Vector{Float64}`:
```julia
# Before
julia> @btime sort!(y, alg=QuickSort) setup=(y = rand(100));
  460.335 ns (0 allocations: 0 bytes)

julia> @btime sort!(y2, alg=QuickSort) setup=(y = rand(100); y2 = ifelse.(rand(length(y)) .< 0.9, y, missing));
  1.861 μs (0 allocations: 0 bytes)

# After
julia> @btime sort!(y, alg=QuickSort) setup=(y = rand(100));
  445.944 ns (0 allocations: 0 bytes)

julia> @btime sort!(y2, alg=QuickSort) setup=(y = rand(100); y2 = ifelse.(rand(length(y)) .< 0.9, y, missing));
  849.537 ns (0 allocations: 0 bytes)
```

Even with this, we're still a bit slower than R (by about 50%) EDIT: for a vector with 10M entries.

Nanosoldier should be run only after merging https://github.com/JuliaLang/julia/pull/27789 since benchmarks only cover the default algorithm, which is currently `MergeSort` with missing values.