@StefanKarpinski 2018-06-27
We may have to do some low down and dirty tricks to get this to match the speed of R's tuned C code.