@nalimilan 2018-06-27
For R it's simpler in that case since `NA` is just a special kind of `NaN`. So basically they should get the same speed as without missing values, while we still have a large penalty.

I've realized my change broke sorting stability, which is visible when calling `sortperm`. AFAICT that's not OK since that method is called even when `alg=MergeSort`, which is documented to be stable. I've not been able to find a simple way of ensuring stability yet. Suggestions welcome.