@nalimilan 2018-06-28
I've found a simple solution: after moving `NaN`s and `missing`s to the end, call `sort!` using the requested algorithm to separate them. That way stability is ensured with `MergeSort` as expected. New timings with 10% of missing values (but no `NaN`s) are quite similar to previous ones.
```julia
julia> @btime sort!(y, alg=QuickSort) setup=(y = rand(100));
  509.596 ns (0 allocations: 0 bytes)

julia> @btime sort!(y2, alg=QuickSort) setup=(y = rand(100); y2 = ifelse.(rand(length(y)) .< 0.9, y, missing));
  881.354 ns (0 allocations: 0 bytes)
```