@nalimilan 2018-07-11
Should we merge this now? It really improves performance for a common case. Maybe we'll be able to find a more general solution (https://github.com/JuliaLang/julia/issues/27781#issuecomment-401110632), but for now this sounds worth it.