@nalimilan 2021-02-26
I've rebased.

Here are the timings
master:
```julia
julia> @btime sort!(y, alg=QuickSort) setup=(y = rand(100));
  463.140 ns (0 allocations: 0 bytes)

julia> @btime sort!(y2, alg=QuickSort) setup=(y = rand(100); y2 = ifelse.(rand(length(y)) .< 0.9, y, missing));
  2.654 μs (0 allocations: 0 bytes)
```

This PR:
```julia
julia> @btime sort!(y, alg=QuickSort) setup=(y = rand(100));
  475.500 ns (0 allocations: 0 bytes)

julia> @btime sort!(y2, alg=QuickSort) setup=(y = rand(100); y2 = ifelse.(rand(length(y)) .< 0.9, y, missing));
  1.411 μs (0 allocations: 0 bytes)
```

Unfortunately it's slower than what I reported above, but I couldn't reproduce these timings, neither on master, 1.5, 1.0 and 0.7, and not even when building the old PR. Not sure how I got them. Anyway the improvement is clear, and for 10M entries we're only 25% slower than R.