@nanosoldier 2021-03-11
Something went wrong when running [your job](https://github.com/JuliaLang/julia/pull/27817#issuecomment-796920956):
```
NanosoldierError: failed to run benchmarks against primary commit: TaskFailedException:
failed process: Process(`git rev-parse HEAD`, ProcessExited(128)) [128]

pipeline_error at ./process.jl:525 [inlined]
read at ./process.jl:412
read at ./process.jl:421
readchomp at ./io.jl:902
build_julia! at /run/media/system/data/nanosoldier/Nanosoldier.jl/src/build.jl:70
build_benchmarksjulia! at /run/media/system/data/nanosoldier/Nanosoldier.jl/src/jobs/BenchmarkJob.jl:299
#17 at ./task.jl:356
```
Logs and partial data can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/4a1ddbb_vs_53f328d/report.md)
cc @christopher-dG