@nanosoldier 2021-03-11
[Your benchmark job](https://github.com/JuliaLang/julia/pull/27817#issuecomment-796940097) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/47b1215_vs_53f328d/report.md). cc @christopher-dG